requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
mysqlclient
django-mysql
djangorestframework
daphne
whitenoise
channels
channels_redis
Pillow